Develop

+++ R/Utilities.R
@@ -1276,7 +1276,7 @@ getPrecOne <- function(x) {
 
 
 setRstoxPrecision <- function(x) {
-    # Get the defines number of digits:
+    # Get the defined number of digits:
     digits <- getRstoxFrameworkDefinitions("digits")
     signifDigits <- getRstoxFrameworkDefinitions("signifDigits")
     
@@ -1304,37 +1304,43 @@ setRstoxPrecision <- function(x) {
             x  <- roundSignif(x, digits = digits, signifDigits = signifDigits)
         }
         else if(getRelevantClass(x) == "sf") {
-            # The precision of geographical coordinates will only be set to the specified digits, with no consideration of significant digits. Using digits = 12 implies a precision of approximately 100 nanometers, which should be enough:
-            precision <- 10^digits
-            
-            # Write the data to a file to utilize the function sf::st_set_precision:
-            outdata <- sf::st_set_precision(x, precision = precision)
-            tmpDir <- file.path(tempdir(), "shapefile")
-            dir.create(tmpDir)
-            tmpFile <- file.path(tmpDir, "nc.shp")
-            
-            # Keep the stratum names and crs: 
-            stratumNames <- outdata$StratumName
-            crs <- sf::st_crs(outdata)
-            
-            # Write the multipolygons to a temporary file (suppress the name abbreviation warning "Field names abbreviated for ESRI Shapefile driver"):
-            suppressWarnings(sf::st_write(outdata, tmpFile, quiet = TRUE))
-            x <- sf::st_read(tmpFile, quiet = TRUE)
-            
-            # Add the stratum names and crs again: 
-            x <- x[, NULL]
-            x$StratumName <- stratumNames
-            # Suppress "replacing crs does not reproject data; use st_transform for that":
-            suppressWarnings(sf::st_crs(x) <- crs)
-            # sf::st_read may read as POLYGON. We want MULTIPOLYGON always:
-            x <- sf::st_cast(x, "MULTIPOLYGON")
-            
-            # Place geometry last, as this seems to be the default in sf, e.g. in st_cast:
-            newOrder <- c(setdiff(names(x), "geometry"), "geometry")
-            x <- x[, newOrder]
-            
-            # delete the shape files:
-            unlink(tmpDir, recursive = TRUE)
+            # Do nothing with waypoints, as the precision is set to 6 digins by DefineSurveyPlan, which is the only datatype that is written to file from sf spation points:
+            if(sf::st_geometry_type(x)[1] == "POINT") {
+                x <- x
+            }
+            else {
+                # The precision of geographical coordinates will only be set to the specified digits, with no consideration of significant digits. Using digits = 12 implies a precision of approximately 100 nanometers, which should be enough:
+                precision <- 10^digits
+                
+                # Write the data to a file to utilize the function sf::st_set_precision:
+                outdata <- sf::st_set_precision(x, precision = precision)
+                tmpDir <- file.path(tempdir(), "shapefile")
+                dir.create(tmpDir)
+                tmpFile <- file.path(tmpDir, "nc.shp")
+                
+                # Keep the stratum names and crs: 
+                stratumNames <- outdata$StratumName
+                crs <- sf::st_crs(outdata)
+                
+                # Write the multipolygons to a temporary file (suppress the name abbreviation warning "Field names abbreviated for ESRI Shapefile driver"):
+                suppressWarnings(sf::st_write(outdata, tmpFile, quiet = TRUE))
+                x <- sf::st_read(tmpFile, quiet = TRUE)
+                
+                # Add the stratum names and crs again: 
+                x <- x[, NULL]
+                x$StratumName <- stratumNames
+                # Suppress "replacing crs does not reproject data; use st_transform for that":
+                suppressWarnings(sf::st_crs(x) <- crs)
+                # sf::st_read may read as POLYGON. We want MULTIPOLYGON always:
+                x <- sf::st_cast(x, "MULTIPOLYGON")
+                
+                # Place geometry last, as this seems to be the default in sf, e.g. in st_cast:
+                newOrder <- c(setdiff(names(x), "geometry"), "geometry")
+                x <- x[, newOrder]
+                
+                # delete the shape files:
+                unlink(tmpDir, recursive = TRUE)
+            }
         }
         # None of the other validOutputDataClasses need setting precision to.
